fix(localstore): step 5 needs to run before all other migrations
