+++ pkg/storer/migration/all_steps.go
@@ -21,6 +21,5 @@ func AllSteps(
 		2: step_02,
 		3: step_03(chunkStore, reserve.ChunkType),
 		4: step_04(sharkyPath, sharkyNoOfShards),
-		5: step_05,
 	}
 }
