+++ pkg/storer/migration/export_test.go
@@ -9,5 +9,4 @@ var (
 	Step_02 = step_02
 	Step_03 = step_03
 	Step_04 = step_04
-	Step_05 = step_05
 )
